+++ projekt Java symulacja/src/package_symulacja/AbsorbtionCoefficientPanel.java
@@ -13,7 +13,7 @@ public class AbsorbtionCoefficientPanel extends JPanel {
 	private JLabel coefficientValueLabel;
 	
 	public AbsorbtionCoefficientPanel() {
-		this.setLayout(new GridLayout(7,1));
+		this.setLayout(new GridLayout(2,1));
 		this.setBorder(BorderFactory.createEmptyBorder(10, 10, 10, 10));
 
 		absorptionCoefficientValue = new JTextField(" ");
